#107 Protect Routes implemented
Created a protected_routes file that prevents navigation to a url if the user has an unauthorized role. Paths are subject to change in future iterations as required.

dashboard_routes updated to use protected_routes.

To Do: in future iteration: Refactor paths to have structure such as
('/parent' +
'/child1'
'/child2'
'/etc')